refactor(www/api-route): replace pgp-request handler with api-route
